move DemuxProcessor and tests two package parents upwards
